pyproject.toml
[project]

name = "Pytodo"
version = "0.2.1"
description = "A todo app in python (API)"
authors = [
    {name = "naughtmarrow",email = "naughtmarrow@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src"},
]


[tool.alembic]

script_location = "%(here)s/migrations"

prepend_sys_path = [
    "."
]
